fix test and bench build problems; add some doc to catalog module
