+++ tests/triesets.rs
@@ -3,7 +3,7 @@ extern crate adapton;
 use std::rc::Rc;
 use adapton::engine::*;
 use adapton::engine::manage::*;
-use adapton::collections::trie::*;
+use adapton::catalog::collections::trie::*;
 
 #[test]
 fn test_is_empty() {
